Fix merge issues and move copy/paste into block pipeline actions hook
