+++ src/devTools/editor/tabs/editTab/editorNodeConfigPanel/EditorNodeConfigPanel.module.scss
@@ -25,6 +25,5 @@
 
 .buttonRow {
   align-items: center;
-  //justify-content: end;
   margin-bottom: 1em;
 }
